requirements.txt
docling>=2.0.0,<3.0.0
docling-core>=2.0.0,<3.0.0
docling-parse>=2.0.0,<3.0.0
docling-ibm-models>=2.0.0,<3.0.0

data-prep-toolkit-transforms[all]>=0.2.2,<1.0.0

# Document format handling
python-pptx>=0.6.21,<1.0.0
reportlab>=4.0.0,<5.0.0
PyPDF2>=3.0.0,<4.0.0
pypdf>=4.0.0,<5.0.0

nltk>=3.8.0,<4.0.0
pandas>=2.0.0,<2.2.0  # Compatible with ibm-watson-machine-learning
numpy>=1.24.0,<2.0.0  # Compatible with thinc and other ML libraries
pyarrow>=14.0.0,<16.0.0

# File handling and utilities
Pillow>=10.0.0,<11.0.0

# Embedding and Vector Storage
sentence-transformers>=2.2.0,<3.0.0
chromadb>=0.4.0,<1.0.0
pymilvus>=2.3.0,<3.0.0
torch>=2.0.0,<3.0.0
transformers>=4.30.0,<5.0.0

# IBM Granite Models
ibm-generative-ai>=3.0.0,<4.0.0
pathlib2>=2.3.7,<3.0.0

pyyaml>=6.0.0,<7.0.0
python-dotenv>=1.0.0,<2.0.0

requests>=2.31.0,<2.32.3  # Compatible with ibm-cos-sdk-core
httpx>=0.25.0,<0.28.0     # Compatible with ollama

tqdm>=4.65.0,<5.0.0

scipy>=1.7.0,<1.14.0      # Compatible with gensim
urllib3>=1.26.0,<2.0.0    # Compatible with ibm-watsonxdata and ibm-cloud-sdk-core

# Development and testing (optional)
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
black>=23.0.0,<25.0.0
flake8>=6.0.0,<8.0.0

ibm-watson==6.0.0
ibm-cloud-sdk-core==3.10.0
ibm-generative-ai==3.0.0
ibm-watsonx-ai==0.1.2